refactor: isolate OCR queue client
